[소스 개선] 소스 세분화
