+++ src/components/Router.js
@@ -5,7 +5,7 @@ import Home from "routes/Home";
 import Navigation from "components/Navigation";
 import Profile from "routes/Profile";
 
-const AppRouter = ({ isLoggedIn, userObj }) => {
+const AppRouter = ({ refreshUser, isLoggedIn, userObj }) => {
     return (
         <>
             <BrowserRouter>
@@ -14,7 +14,7 @@ const AppRouter = ({ isLoggedIn, userObj }) => {
                     {isLoggedIn ? (
                         <>
                             <Route path="/" element={<Home userObj={userObj} />} />
-                            <Route path="/profile" element={<Profile userObj={userObj} />} />
+                            <Route path="/profile" element={<Profile userObj={userObj} refreshUser={refreshUser} />} />
                         </>
                     ) : (
                         <Route path="/" element={<Auth />} />
