+++ src/routes/Auth.js
@@ -1,38 +1,10 @@
+import React from "react";
 import { authService } from "fBase";
-import { getAuth, signInWithEmailAndPassword, createUserWithEmailAndPassword, GoogleAuthProvider, GithubAuthProvider, signInWithPopup } from "firebase/auth";
-import React, { useState } from "react";
+import {GoogleAuthProvider, GithubAuthProvider, signInWithPopup } from "firebase/auth";
+import AuthForm from "components/AuthForm";
 
 const Auth = () => {
-    const [email, setEmail] = useState("");
-    const [password, setPassword] = useState("");
-    const [newAccount, setNewAccount] = useState(false);
-    const [error, setError] = useState("");
-    const onChange = (event) => {
-        const {
-            target: { name, value },
-        } = event;
-        if (name === "email") {
-            setEmail(value);
-        } else if (name === "password") {
-            setPassword(value);
-        }
-    };
-    const onSubmit = async (event) => {
-        event.preventDefault();
-        try {
-            let data;
-            const auth = getAuth();
-            if (newAccount) {
-                data = await createUserWithEmailAndPassword(auth, email, password);
-            } else {
-                data = await signInWithEmailAndPassword(auth, email, password);
-            }
-            console.log(data);
-        } catch (error) {
-            setError(error.message);
-        }
-    };
-    const toggleAccount = () => setNewAccount((prev) => !prev);
+    
     const onSocialClick = async (e) => {
         const {
             target: { name },
@@ -47,13 +19,7 @@ const Auth = () => {
     };
     return (
         <div>
-            <form onSubmit={onSubmit}>
-                <input name="email" type="email" placeholder="Email" required value={email} onChange={onChange} />
-                <input name="password" type="Password" placeholder="Password" required value={password} onChange={onChange} />
-                <input type="submit" value={newAccount ? "Create New Account" : "Log In"} />
-            </form>
-            {error}
-            <span onClick={toggleAccount}>{newAccount ? "Sign In" : "Create Account"}</span>
+            <AuthForm />
             <div>
                 <button name="google" onClick={onSocialClick}>
                     Continue With Google
